prediction-api/requirements.txt
flask==3.0.0
gunicorn==21.2.0
pandas==1.5.3
scikit-learn==1.3.2
xgboost==2.0.2
google-cloud-storage==2.13.0
numpy==1.23.2